+++ VectorField.cpp
@@ -13,14 +13,14 @@ VectorField::~VectorField() {
 	}
 }
 
-void VectorField::setValue(glm::vec3 value, uint index) {
+void VectorField::setValue(glm::dvec3 value, uint index) {
 	components[0]->setValue(value.x, index);
 	components[1]->setValue(value.y, index);
 	components[2]->setValue(value.z, index);
 }
 
-glm::vec3 VectorField::getValue(uint index) const {
-	return glm::vec3(
+glm::dvec3 VectorField::getValue(uint index) const {
+	return glm::dvec3(
 		components[0]->getValue(index),
 		components[1]->getValue(index),
 		components[2]->getValue(index)
@@ -38,7 +38,7 @@ void VectorField::write(std::string path, bool header) const {
     }
     // Write values
     for (uint i = 0; i < samples; ++i) {
-		const glm::vec3 v = getValue(i);
+		const glm::dvec3 v = getValue(i);
         file << std::setprecision(DPRECIS) << v.x << " "
             << std::setprecision(DPRECIS) << v.y << " "
             << std::setprecision(DPRECIS) << v.z << std::endl;
@@ -55,7 +55,7 @@ void VectorField::write2d(std::string path) const {
 
     // Write verts
     for (uint i = 0; i < samples; ++i) {
-		const glm::vec3 v = getValue(i);
+		const glm::dvec3 v = getValue(i);
         file << std::setprecision(DPRECIS) << v.x << " "
             << std::setprecision(DPRECIS) << v.y << std::endl;
     }
